refactor: rename file plugingloader.ts to pluging-loader.ts
