+++ packages/opencensus-nodejs/src/index.ts
@@ -19,4 +19,4 @@ import {Tracing} from './trace/tracing';
 
 const tracing: types.Tracing = Tracing.instance;
 
-export {tracing};
+export = tracing;
