--- packages/opencensus-nodejs/tsconfig.full.json
@@ -1,17 +1,0 @@
-{
-  "extends": "./node_modules/gts/tsconfig-google.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "build",
-    "target": "es6",
-    "allowUnreachableCode": true,
-    "noUnusedParameters": false,
-    "noImplicitAny": false,
-    "noImplicitThis": false,
-    "noEmitOnError": false
-  },
-"include": [
-  "src/**/*.ts",
-  "test/**/*.ts"
- ]
-}
