+++ packages/opencensus-nodejs/tsconfig.json
@@ -15,8 +15,8 @@
     "sourceMap": false
   },
   "include": [
-    "src/*.ts",
-    "test/*.ts",
+    "src/**/*.ts",
+    "test/**/*.ts"
   ],
   "exclude": [
     "node_modules"
